Refactored location of weightmap trainer files to be consistent with sparse trainer #7
